Design: 레이아웃 UI 변경
