+++ src/app/globals.css
@@ -3,7 +3,7 @@
 @tailwind utilities;
 
 :root {
-  --font-pretendard: "Pretendard";
+  --font-pretendard: 'Pretendard';
 }
 
 body {
